Mergin with PC

+++ backend/crokiclothes/views.py
@@ -1,8 +1,5 @@
-
-import base64
-import io
+import os
 from django.shortcuts import render
-import pika
 import requests
 # Create your views here.
 from rest_framework import viewsets, status
@@ -10,7 +7,7 @@ from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework.parsers import MultiPartParser, FormParser
 from rest_framework.permissions import *
-from .producer import publish
+
 
 
 from .serializers import *
@@ -246,11 +243,20 @@ class ConvertingImagesForAugObject(APIView):
         aug_model_id = request.data.get('aug_model')
         convering_image_id = request.data.get('id')
         converting_image = ImagesToConvert.objects.get(id=convering_image_id)
-        converting_image.delete()
-        rest_converting_images = ImagesToConvert.objects.all().filter(aug_model=aug_model_id)
-        serializer = ConvertingImagesSerializer(rest_converting_images,many=True)
+        converting_image_path = str(converting_image.convertingimages.path)
+        print(converting_image_path)
+        if os.path.exists(converting_image_path):
+            os.remove(converting_image_path)
+            converting_image.delete()
+            rest_converting_images = ImagesToConvert.objects.all().filter(aug_model=aug_model_id)
+            serializer = ConvertingImagesSerializer(rest_converting_images,many=True)
+    
+            return Response(serializer.data)
+        else:
+            print("File doens't exist")
+            return Response("File does not exist")
         
-        return Response(serializer.data)
+            
 
 class AugmentedObjView(APIView):
     parser_classes = (MultiPartParser, FormParser)
@@ -275,6 +281,22 @@ class AugmentedObjView(APIView):
         # aug_object_serialized = ArObjectSerializer(aug_object,many=True)
         # return Response(aug_object_serialized.data, status=status.HTTP_200_OK)
         return Response("None")
+    
+    def post(self, request, pk=None):
+        # url = 'http://192.168.0.108:5000/get_mesh' ## PC  
+        # server = requests.get(url)
+        # print(server.data)
+        print("Has recieved smth")
+        # print(request.data)
+        mesh_obj = request.data.get('texturedMesh.obj')
+        print(mesh_obj)
+        # serializer = ArObjectSerializer(data=mesh_obj)
+        # if (serializer.is_valid(raise_exception=True)):
+        #     serializer.save()
+        #     print("instance saved")
+        # else:    
+        #     print("failed to save object")
+        return Response("recieved")
     
 
     
@@ -324,13 +346,12 @@ class WorkshopView(APIView):
 class ConvertingView(APIView):
     parser_classes = (MultiPartParser, FormParser)
     permission_classes = (IsAdminUser, )
+    print("Sending images method has started")
     
     def post(self, request, pk=None):
         folder_name = request.data.get('folder_name')
         try:
             aug_model = ArObject.objects.get(cloth=pk)
-            
-            print(aug_model.id)
             try:
                 images_to_convert = ImagesToConvert.objects.all().filter(aug_model=aug_model.id)
                 image_files = []
@@ -338,11 +359,14 @@ class ConvertingView(APIView):
                     image_files.append(('files', (img.convertingimages.name, img.convertingimages.file)))
 
                 
-                url = 'http://192.168.1.53:5000/bulkUpload'
+                # url = 'http://192.168.0.101:5000/bulkUpload' ## Laptop
+                url = 'http://192.168.0.108:5000/bulkUpload' ## PC
+
+                
                 form_data = {'foldername':folder_name}
                 server = requests.post(url, data=form_data, files=image_files)
                 output = server.text
-                print(output)
+                # print(output)
                 
                 return Response(output, status=status.HTTP_200_OK)
